title screen fully implemented plus sml chngs
title screen has been fully implemented with some small fixes with typing and indentations

+++ Space Invaders/RaylibStarterCPP/Enemies.h
@@ -6,7 +6,7 @@ public:
 	Enemies();
 	~Enemies();
 
-	Enemies(int x, int y, int size);
+	Enemies(int x, int y, float size);
 
 	void update(Vector2 updatePos);
 	void draw() override;
